refactor(cjk): split the two files this branch pushed over the size limit, and stop applyTransform reading the store
Everything `check:predelta` found in one pass, fixed rather than baselined.

`formatter.ts` reached 355 lines. Table recognition and cell-wise formatting is
a coherent concern — the delimiter row is structure that must survive verbatim,
a pipe inside inline code is not a cell boundary, and a cell's padding is the
author's alignment — so it moves to `formatterTables.ts` (182 + 202). The cell
formatter is INJECTED rather than imported, so the two modules do not depend on
each other: the pipeline drives the tables, not the other way round.

`markdownParser.ts` had grown 321 to 375, past its frozen baseline. Its two
LINE-ORIENTED detectors (indented code, reference sections) need a line-by-line
walk rather than a regex and are the two longest by a wide margin; they move to
`markdownParserBlocks.ts`. They append to the caller's array in place, exactly
as they did inline, because the detectors are order-dependent. That put the
parser at 284 — under the 300 limit — so its baseline entry is DELETED rather
than lowered. `rules.test.ts` is back under its cap too, by compressing comments
that duplicate rationale stated in full elsewhere.

Splitting the parser made it circular with its new sibling: each needed
`ProtectedRegion`. The types move to `types.ts`, which already owns this
package's contracts, and the reason is recorded there so a future reader does
not move them back.

`applyTransform` took its settings from `useSettingsStore.getState()` and
`shouldPreserveTwoSpaceBreaks()`, which made a pure string-to-string function
depend on two singletons and forced its tests to mock the app's SETTINGS STORE
— mocking app state rather than a boundary, which `lint:mock-boundaries`
correctly refuses. They are parameters now, and `currentTransformSettings()`
does the reads in one named place the handler calls once. Both new test files
drop their store mocks, and `sourceCjkActions.test.ts`'s settings-store entry is
removed from the baseline to record the win from WI-CJKF1.3.

`FormatResult` left the barrel (re-exported, never imported as a type) and
`isInsideRegion` is module-private in the new tables file — both were new dead
exports that `lint:knip-baseline` caught.

One real defect surfaced here that the test suite could not:
`formatMarkdownChecked` kept its `: string` return annotation through the rename
and returned an object from it. Every test passed, because vitest does not
typecheck. `pnpm typecheck` is what found it.

Plan: dev-docs/plans/20260821-cjk-formatter-correctness.md
