+++ scripts/file-size-baseline.json
@@ -45,7 +45,6 @@
     "src/export/useExportOperations.ts": 446,
     "src/hooks/useTabDragOut.ts": 356,
     "src/hooks/useUpdateChecker.ts": 407,
-    "src/lib/cjkFormatter/markdownParser.ts": 321,
     "src/lib/cjkFormatter/quotePairing.ts": 378,
     "src/lib/ghaWorkflow/completion/expressionCompletion.ts": 422,
     "src/lib/ghaWorkflow/cron/readable.ts": 386,
@@ -116,7 +115,7 @@
     "src/hooks/useTabDragOut.test.ts": 1150,
     "src/hooks/useUpdateChecker.comprehensive.test.ts": 924,
     "src/hooks/useUpdateChecker.test.ts": 932,
-    "src/lib/cjkFormatter/rules.test.ts": 1865,
+    "src/lib/cjkFormatter/rules.test.ts": 1863,
     "src/pages/settings/components.test.tsx": 989,
     "src/plugins/aiSuggestion/__tests__/tiptap.test.ts": 1761,
     "src/plugins/autoPair/__tests__/handlersExtra.test.ts": 868,
